Initial commit ==============
    2015-10-29

Contains two modules; settings and webnote. These are files copied and
adapted from the code written for the P3 programme.

This initial coimmit is halfway through a refactoring of the existing
code, invilving throwing out a lot of proceduaral code in favour of
declaratives, snd heavy use of the property keyword, use of which
permits the coding of object attributes which are computed by methods.

This design has the great benefit of making all the computed values
(including lists of related pages) available to the user in
object.attribute format, and it only computes values when they are
asked for.

There is still a lot of legacy code in the webnote module, which is
preserved here for now, as some of it refers to computing things like
siblings and children, and may yet prove useful.