Page development
Sorted out a lot of empty property methods, which are intended to
supply values for things like the lists of siblings and children, the
parent document.

Wrote the methods which find the parent and paired directories, and
tested these at various levels in the filesystem. They seem to work,
and now I have a webnote model of each filespace. Finds the filename
for the addressed resource reliably now as well.