Nearing deployable completion =============================
I have created and completed a Directory class, which will parse a
directrory listing into file types, and has a series of property
attributes to permit accessing the various lists.

I have also written enough code to the Page class to deploy it
somewhere. This is, it reads the address, finds the file that address
points at, and generates lists of psiblings and children, as well as
documents and such associated with that page (in its paired directory).

There is some work to do around the next and previous links, and some
testing, which would best be done in a deployment.