Bug bashing following adding in page objects on Page template.
Had worked some methods in the Page class to provide lists of siblings
and children as Page objects, instead of (actually, in addition to)
the lists of (link, text) tuples. This caused the system of saving new
pages to fail, so I had to track down that failure and fix it.